updated googleCalendar variable
